adds untested BarrelPivot movement command
This commit adds untested BarrelPivot code which uses the absolute
encoder to move the arm based on joystick inputs.

SetZeroAsCurrentPosition will be helpful for testing, but shouldn't be
needed once we determine the permanent "zero".

Inversion of the motor/absolute encoder (so that "up" is upward
rotation and "down" is downward rotation) is currently unknown.