adds an "Enabled" switch to each Testing Dashboard tab
Each registered subsystem includes a "TestingDashboard/Enabled"
element. The default is true (Enabled). If the switch is turned off,
the subsystem will stop posting new values to the dashboard. This may
be helpful in competition when we want to limit our bandwidth usage.

This also changes how TDBoolean is registered (it was
a number before).