renames 'resetTargetAngle.java' to 'ResetTargetAngle.java'
The name of a java file must match the name of the contained class
(the capitalization of the class name changed recently). Is this a
Windows thing that java can't recognize the capitalization of a file?
It doesn't compile in my non-Windows environment. (Also, VSCode didn't
recognize the filename change as a git-worthy event. I had to use
'git mv')