updates autos
This commit updates the current autos that the robot has. Rather than
keeping the note stuck in the Intake, it is moved into the Barrel which
subsequently allows for faster shooting. To combat the robot drifting,
the P for the translation & rotation PID was reduced from 5.0 to 1.0.
This commit also sends the AutoChooser to SmartDashboard, so that it
appears on the Driver Station.

Test the PID again with access to AprilTags at WPI***