updates autos
This commit makes several updates to the autos. It renames all the autos
so that they are less confusing by changing startTop to startAmp,
startBot to startSource, and closeAutoBot to closeAutoStage. There
were a few changes made to the pathing, including going at the note
closest to the amp straight instead of rotating to it. It fixes most of
the messy competition changes that were made in a hurry.